Matlab over batch and exit values
Execute Matlab script over batch because matlab does not find the script
despite changing the working directory of the process. Additionaly the
thread stops if a process returns an exit value unequal to zero (error).
This happend because some material values in the csv files were not in
range for GPR Max software.